Move action list management into a separate class and module.
registerAction and cancelAction remain in DeviceTree because their
purpose is wholly specific to the device tree. The action list has
no concept of registering or canceling actions -- there is only
append and remove.
